AshGeo.Postgis (AshGeo v0.3.0)

PostGIS functions for use with Ash.Expr

Functions

gis_index(column, name \\ nil)
(since 0.1.0) (macro)

Create

st_area(geometry)
(since 0.1.0) (macro)

st_as_binary(geometry)
(since 0.1.0) (macro)

st_as_text(geometry)
(since 0.1.0) (macro)

st_bd_m_poly_from_text(wkt, srid)
(since 0.1.0) (macro)

st_bd_poly_from_text(wkt, srid)
(since 0.1.0) (macro)

st_boundary(geometry)
(since 0.1.0) (macro)

st_buffer(geometry, double)
(since 0.1.0) (macro)

st_buffer(geometry, double, integer)
(since 0.1.0) (macro)

st_build_area(geometryA)
(since 0.1.0) (macro)

st_centroid(geometry)
(since 0.1.0) (macro)

st_collect(geometryList)
(since 0.1.0) (macro)

st_collect(geometryA, geometryB)
(since 0.1.0) (macro)

st_contains(geometryA, geometryB)
(since 0.1.0) (macro)

st_convex_hull(geometry)
(since 0.1.0) (macro)

st_covered_by(geometryA, geometryB)
(since 0.1.0) (macro)

st_covers(geometryA, geometryB)
(since 0.1.0) (macro)

st_crosses(geometryA, geometryB)
(since 0.1.0) (macro)

st_difference(geometryA, geometryB)
(since 0.1.0) (macro)

st_dimension(geometry)
(since 0.1.0) (macro)

st_disjoint(geometryA, geometryB)
(since 0.1.0) (macro)

st_distance(geometryA, geometryB)
(since 0.1.0) (macro)

st_distance_in_meters(geometryA, geometryB)
(since 0.1.0) (macro)

Casts the 2 geometries given to geographies in order to return distance in meters.

st_distance_sphere(geometryA, geometryB)
(since 0.1.0) (macro)

Please note that ST_Distance_Sphere has been deprecated as of Postgis 2.2. Postgis 2.1 is no longer supported on PostgreSQL >= 9.5. This macro is still in place to support users of PostgreSQL <= 9.4.x.

st_distancesphere(geometryA, geometryB)
(since 0.1.0) (macro)

st_dwithin(geometryA, geometryB, float)
(since 0.1.0) (macro)

st_dwithin_in_meters(geometryA, geometryB, float)
(since 0.1.0) (macro)

Casts the 2 geometries given to geographies in order to check for distance in meters.

st_end_point(geometry)
(since 0.1.0) (macro)

st_envelope(geometry)
(since 0.1.0) (macro)

st_equals(geometryA, geometryB)
(since 0.1.0) (macro)

st_extent(geometry)
(since 0.1.0) (macro)

st_exterior_ring(geometry)
(since 0.1.0) (macro)

st_flip_coordinates(geometryA)
(since 0.1.0) (macro)

st_generate_points(geometryA, npoints)
(since 0.1.0) (macro)

st_generate_points(geometryA, npoints, seed)
(since 0.1.0) (macro)

st_geom_coll_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_geom_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_geometry_n(geometry, int)
(since 0.1.0) (macro)

st_geometry_type(geometry)
(since 0.1.0) (macro)

st_interior_ring_n(geometry, int)
(since 0.1.0) (macro)

st_intersection(geometryA, geometryB)
(since 0.1.0) (macro)

st_intersects(geometryA, geometryB)
(since 0.1.0) (macro)

st_is_closed(geometry)
(since 0.1.0) (macro)

st_is_ring(geometry)
(since 0.1.0) (macro)

st_is_simple(geometry)
(since 0.1.0) (macro)

st_length(geometry)
(since 0.1.0) (macro)

st_line_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_line_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_linestring_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_linestring_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_m(geometry)
(since 0.1.0) (macro)

st_m_geom_coll_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_m_geom_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_m_geometry_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_m_line_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_m_line_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_m_point_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_m_point_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_m_poly_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_m_poly_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_make_box_2d(geometryA, geometryB)
(since 0.1.0) (macro)

st_mem_union(geometryList)
(since 0.1.0) (macro)

st_num_geometries(geometry)
(since 0.1.0) (macro)

st_num_interior_ring(geometry)
(since 0.1.0) (macro)

st_num_interior_rings(geometry)
(since 0.1.0) (macro)

st_num_points(geometry)
(since 0.1.0) (macro)

st_overlaps(geometryA, geometryB)
(since 0.1.0) (macro)

st_point(x, y)
(since 0.1.0) (macro)

st_point_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_point_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_point_n(geometry, int)
(since 0.1.0) (macro)

st_point_on_surface(geometry)
(since 0.1.0) (macro)

st_poly_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_polygon_from_text(text, srid \\ -1)
(since 0.1.0) (macro)

st_polygon_from_wkb(bytea, srid \\ -1)
(since 0.1.0) (macro)

st_relate(geometryA, geometryB)
(since 0.1.0) (macro)

st_relate(geometryA, geometryB, intersectionPatternMatrix)
(since 0.1.0) (macro)

st_set_srid(geometry, srid)
(since 0.1.0) (macro)

st_shift_longitude(geometry)
(since 0.1.0) (macro)

st_srid(geometry)
(since 0.1.0) (macro)

st_start_point(geometry)
(since 0.1.0) (macro)

st_sym_difference(geometryA, geometryB)
(since 0.1.0) (macro)

st_touches(geometryA, geometryB)
(since 0.1.0) (macro)

st_transform(wkt, srid)
(since 0.1.0) (macro)

st_union(geometryList)
(since 0.1.0) (macro)

st_union(geometryA, geometryB)
(since 0.1.0) (macro)

st_within(geometryA, geometryB)
(since 0.1.0) (macro)

st_x(geometry)
(since 0.1.0) (macro)

st_y(geometry)
(since 0.1.0) (macro)

st_z(geometry)
(since 0.1.0) (macro)